move icons to top of autocomplete
