+++ datahub-web-react/src/app/search/autoComplete/AutoCompleteEntityIcon.tsx
@@ -7,43 +7,26 @@ import { useEntityRegistry } from '../../useEntityRegistry';
 import { IconStyleType } from '../../entity/Entity';
 
 const PreviewImage = styled(Image)`
-    width: auto;
+    height: 16px;
+    width: 16px;
     object-fit: contain;
     background-color: transparent;
 `;
 
 type Props = {
     entity: Entity;
-    scale: number;
 };
 
-const DEFAULT_ICON_SIZE = 12;
-const DEFAULT_PREVIEW_IMAGE_SIZE = 22;
-
-const AutoCompleteEntityIcon = ({ entity, scale }: Props) => {
+const AutoCompleteEntityIcon = ({ entity }: Props) => {
     const entityRegistry = useEntityRegistry();
-
-    const iconFontSize = Math.floor(DEFAULT_ICON_SIZE * scale);
-    const previewImageSize = Math.floor(DEFAULT_PREVIEW_IMAGE_SIZE * scale);
 
     const genericEntityProps = entityRegistry.getGenericEntityProperties(entity.type, entity);
     const platformLogoUrl = genericEntityProps?.platform?.properties?.logoUrl;
     const platformName = getPlatformName(genericEntityProps);
-    const icon =
-        (platformLogoUrl && (
-            <PreviewImage
-                preview={false}
-                src={platformLogoUrl}
-                alt={platformName || ''}
-                style={{
-                    width: previewImageSize,
-                    height: previewImageSize,
-                }}
-            />
-        )) ||
-        entityRegistry.getIcon(entity.type, iconFontSize, IconStyleType.ACCENT);
-
-    return icon;
+    return (
+        (platformLogoUrl && <PreviewImage preview={false} src={platformLogoUrl} alt={platformName || ''} />) ||
+        entityRegistry.getIcon(entity.type, 12, IconStyleType.ACCENT)
+    );
 };
 
 export default AutoCompleteEntityIcon;
